Refactor OrderUpdater to decorate Spree's one
This replaces the nasty class eval and provides some encapsulation
